traffic light based on time due added, Time needs fixing to NZ date and time

+++ BackEnd/ConsoleUser/Models/Domain/DashboardTicketData.cs
@@ -16,7 +16,7 @@ namespace ConsoleUser.Models.Domain
         [AllowNull] public string Type { get; set; }
         [AllowNull] public string url { get; set; }
         [AllowNull] public string TrafficLight { get; set; }
-        [AllowNull] public int SortPriority { get; set; }
+        [AllowNull] public double SortPriority { get; set; }
 
         // trying Enums
         //private Priority _priority;
